unify one more arg

+++ test/conformance/api/v1/revision_timeout_test.go
@@ -47,7 +47,7 @@ func createService(t *testing.T, clients *test.Clients, names test.ResourceNames
 	return clients.ServingClient.Services.Create(service)
 }
 
-func updateServiceWithTimeout(clients *test.Clients, names test.ResourceNames, revisionTimeoutSeconds int) error {
+func updateServiceWithTimeout(clients *test.Clients, names test.ResourceNames, revisionTimeoutSeconds int64) error {
 	patches := []jsonpatch.JsonPatchOperation{{
 		Operation: "replace",
 		Path:      "/spec/template/spec/timeoutSeconds",
