+++ test/conformance/api/v1alpha1/revision_timeout_test.go
@@ -48,7 +48,7 @@ func createLatestService(t *testing.T, clients *test.Clients, names test.Resourc
 	return clients.ServingAlphaClient.Services.Create(service)
 }
 
-func updateServiceWithTimeout(clients *test.Clients, names test.ResourceNames, revisionTimeoutSeconds int) error {
+func updateServiceWithTimeout(clients *test.Clients, names test.ResourceNames, revisionTimeoutSeconds int64) error {
 	patches := []jsonpatch.JsonPatchOperation{{
 		Operation: "replace",
 		Path:      "/spec/template/spec/timeoutSeconds",
